+++ packages/abc/onboarding/onboarding.service.ts
@@ -1,3 +1,4 @@
+import { Directionality } from '@angular/cdk/bidi';
 import { DOCUMENT } from '@angular/common';
 import {
   ApplicationRef,
@@ -8,9 +9,11 @@ import {
   Injectable,
   Injector,
   OnDestroy,
+  Optional,
 } from '@angular/core';
 import { Router } from '@angular/router';
 import { DelonLocaleService } from '@delon/theme';
+import { AlainConfigService } from '@delon/util';
 import { of, pipe, Subscription } from 'rxjs';
 import { delay, switchMap } from 'rxjs/operators';
 import { OnboardingComponent } from './onboarding.component';
@@ -30,6 +33,11 @@ export class OnboardingService implements OnDestroy {
     return this.doc;
   }
 
+  /**
+   * Get whether it is booting
+   *
+   * 获取是否正在引导中
+   */
   get running(): boolean {
     return this._running;
   }
@@ -41,6 +49,8 @@ export class OnboardingService implements OnDestroy {
     private router: Router,
     private injector: Injector,
     @Inject(DOCUMENT) private doc: any,
+    private configSrv: AlainConfigService,
+    @Optional() private directionality: Directionality,
   ) {}
 
   private attach(): void {
@@ -101,7 +111,8 @@ export class OnboardingService implements OnDestroy {
       ...this.i18n.getData('onboarding'),
       ...items[this.active],
     } as OnboardingItem;
-    Object.assign(this.compRef.instance, { item, config: this.config, active: this.active, max: items.length });
+    const dir = this.configSrv.get('onboarding')!.direction || this.directionality.value;
+    Object.assign(this.compRef.instance, { item, config: this.config, active: this.active, max: items.length, dir });
     const pipes = [
       switchMap(() => (item.url ? this.router.navigateByUrl(item.url) : of(true))),
       switchMap(() => {
@@ -123,6 +134,11 @@ export class OnboardingService implements OnDestroy {
       );
   }
 
+  /**
+   * Start a new user guidance
+   *
+   * 开启新的用户引导流程
+   */
   start(config: OnboardingConfig): void {
     if (this.running) {
       return;
@@ -141,6 +157,11 @@ export class OnboardingService implements OnDestroy {
     this.showItem(true);
   }
 
+  /**
+   * Next
+   *
+   * 下一步
+   */
   next(): void {
     if (this._running || this.active + 1 >= this.config.items!.length) {
       this.done();
@@ -151,6 +172,11 @@ export class OnboardingService implements OnDestroy {
     this.showItem();
   }
 
+  /**
+   * Prev
+   *
+   * 上一步
+   */
   prev(): void {
     if (this._running || this.active - 1 < 0) {
       return;
@@ -160,6 +186,11 @@ export class OnboardingService implements OnDestroy {
     this.showItem();
   }
 
+  /**
+   * Done
+   *
+   * 完成
+   */
   done(): void {
     this.type = 'done';
     this.destroy();