+++ packages/schematics/application/index.ts
@@ -45,14 +45,17 @@ function removeOrginalFiles(): (host: Tree) => void {
       `${project.root}/README.md`,
       `${project.root}/tslint.json`,
       `${project.sourceRoot}/main.ts`,
+      `${project.sourceRoot}/test.ts`,
       `${project.sourceRoot}/environments/environment.prod.ts`,
       `${project.sourceRoot}/environments/environment.ts`,
       `${project.sourceRoot}/styles.less`,
+      `${project.sourceRoot}/favicon.ico`,
       `${project.sourceRoot}/app/app.module.ts`,
       `${project.sourceRoot}/app/app.component.spec.ts`,
       `${project.sourceRoot}/app/app.component.ts`,
       `${project.sourceRoot}/app/app.component.html`,
       `${project.sourceRoot}/app/app.component.less`,
+      `${project.sourceRoot}/app/app-routing.module.ts`,
     ]
       .filter(p => host.exists(p))
       .forEach(p => host.delete(p));
@@ -63,9 +66,22 @@ function fixAngularJson(options: ApplicationOptions): (host: Tree) => void {
   return (host: Tree) => {
     const json = getAngular(host);
     const _project = getProjectFromWorkspace(json, options.project);
-
+    const architect = (_project.targets || _project.architect)!;
     // Add proxy.conf.json
-    (_project.targets || _project.architect)!.serve!.options.proxyConfig = 'proxy.conf.json';
+    architect.serve!.options.proxyConfig = 'proxy.conf.json';
+    // 调整budgets
+    const budgets = architect.build.configurations.production.budgets as Array<{
+      type: string;
+      maximumWarning: string;
+      maximumError: string;
+    }>;
+    if (budgets && budgets.length > 0) {
+      const initial = budgets.find(w => w.type === 'initial');
+      if (initial) {
+        initial.maximumWarning = '2mb';
+        initial.maximumError = '3mb';
+      }
+    }
 
     overwriteAngular(host, json);
     return host;
@@ -96,6 +112,7 @@ function addDependenciesToPackageJson(options: ApplicationOptions): (host: Tree)
         `ng-alain@${VERSION}`,
         `ng-alain-codelyzer@DEP-0.0.0-PLACEHOLDER`,
         `ng-alain-plugin-theme@DEP-0.0.0-PLACEHOLDER`,
+        `source-map-explorer@DEP-0.0.0-PLACEHOLDER`,
         `@delon/testing@${VERSION}`,
       ],
       'devDependencies',
@@ -115,10 +132,12 @@ function addRunScriptToPackageJson(): (host: Tree) => void {
   return (host: Tree) => {
     const json = getPackage(host, 'scripts');
     if (json == null) return host;
+    json.scripts['ng-high-memory'] = `node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng`;
     json.scripts.start = `ng s -o`;
     json.scripts.hmr = `ng s -o --hmr`;
-    json.scripts.build = `node --max_old_space_size=5120 ./node_modules/@angular/cli/bin/ng build --prod`;
-    json.scripts.analyze = `node --max_old_space_size=5120 ./node_modules/@angular/cli/bin/ng build --prod --stats-json`;
+    json.scripts.build = `npm run ng-high-memory build -- --prod`;
+    json.scripts.analyze = `npm run ng-high-memory build -- --prod --source-map`;
+    json.scripts['analyze:view'] = `source-map-explorer dist/**/*.js`;
     json.scripts['test-coverage'] = `ng test --code-coverage --watch=false`;
     json.scripts['color-less'] = `ng-alain-plugin-theme -t=colorLess`;
     json.scripts.theme = `ng-alain-plugin-theme -t=themeCss`;
@@ -158,7 +177,6 @@ function addCodeStylesToPackageJson(): (host: Tree) => Tree {
       host,
       [
         `tslint-config-prettier@DEP-0.0.0-PLACEHOLDER`,
-        `tslint-language-service@DEP-0.0.0-PLACEHOLDER`,
         `pretty-quick@DEP-0.0.0-PLACEHOLDER`,
         `husky@DEP-0.0.0-PLACEHOLDER`,
         `prettier@DEP-0.0.0-PLACEHOLDER`,
@@ -379,9 +397,14 @@ function fixVsCode(): (host: Tree) => void {
   };
 }
 
-function finished(): (_host: Tree, context: SchematicContext) => void {
+function install(): (_host: Tree, context: SchematicContext) => void {
   return (_host: Tree, context: SchematicContext) => {
     context.addTask(new NodePackageInstallTask());
+  };
+}
+
+function finished(): (_host: Tree, context: SchematicContext) => void {
+  return (_host: Tree, _context: SchematicContext) => {
     spinner.succeed(`Congratulations, NG-ALAIN scaffold generation complete.`);
   };
 }
@@ -408,6 +431,7 @@ export default function (options: ApplicationOptions): Rule {
       fixLang(options),
       fixVsCode(),
       fixAngularJson(options),
+      install(),
       finished(),
     ])(host, context);
   };