+++ packages/util/src/config/abc/index.ts
@@ -12,3 +12,5 @@ export * from './st.type';
 export * from './xlsx.type';
 export * from './zip.type';
 export * from './media.type';
+export * from './pdf.type';
+export * from './onboarding.type';