+++ packages/util/src/config/config.types.ts
@@ -6,7 +6,9 @@ import {
   AlainLoadingConfig,
   AlainLodopConfig,
   AlainMediaConfig,
+  AlainOnboardingConfig,
   AlainPageHeaderConfig,
+  AlainPdfConfig,
   AlainQRConfig,
   AlainSEConfig,
   AlainSGConfig,
@@ -29,6 +31,7 @@ export interface AlainConfig {
   errorCollect?: AlainErrorCollectConfig;
   image?: AlainImageConfig;
   loading?: AlainLoadingConfig;
+  onboarding?: AlainOnboardingConfig;
   lodop?: AlainLodopConfig;
   pageHeader?: AlainPageHeaderConfig;
   qr?: AlainQRConfig;
@@ -39,6 +42,7 @@ export interface AlainConfig {
   sf?: AlainSFConfig;
   xlsx?: AlainXlsxConfig;
   zip?: AlainZipConfig;
+  pdf?: AlainPdfConfig;
   media?: AlainMediaConfig;
   acl?: AlainACLConfig;
   auth?: AlainAuthConfig;