+++ scripts/ci/build-delon.sh
@@ -5,6 +5,8 @@ set -u -e -o pipefail
 
 cd $(dirname $0)/../..
 
+source ./scripts/ci/utils.sh
+
 DEBUG=false
 PACKAGES=(util
   testing
@@ -38,23 +40,6 @@ buildLess() {
   node ./scripts/build/generate-css.js
 }
 
-containsElement () {
-  local e
-  for e in "${@:2}"; do [[ "$e" == "$1" ]] && return 0; done
-  return 1
-}
-
-updateVersionReferences() {
-  NPM_DIR="$1"
-  (
-    echo "======    VERSION: Updating version references in ${NPM_DIR}"
-    cd ${NPM_DIR}
-    perl -p -i -e "s/ZORRO\-0\.0\.0\-PLACEHOLDER/${ZORROVERSION}/g" $(grep -ril ZORRO\-0\.0\.0\-PLACEHOLDER .) < /dev/null 2> /dev/null
-    perl -p -i -e "s/PEER\-0\.0\.0\-PLACEHOLDER/^${VERSION}/g" $(grep -ril PEER\-0\.0\.0\-PLACEHOLDER .) < /dev/null 2> /dev/null
-    perl -p -i -e "s/0\.0\.0\-PLACEHOLDER/${VERSION}/g" $(grep -ril 0\.0\.0\-PLACEHOLDER .) < /dev/null 2> /dev/null
-  )
-}
-
 addBanners() {
   for file in ${1}/*; do
     if [[ -f ${file} && "${file##*.}" != "map" ]]; then
@@ -69,8 +54,6 @@ copySchemas() {
   cp ${SOURCE}/abc/onboarding/schema.json ${DIST}/abc/onboarding/schema.json
 }
 
-VERSION=$(node -p "require('./package.json').version")
-ZORROVERSION=$(node -p "require('./package.json').dependencies['ng-zorro-antd']")
 echo "=====BUILDING: Version ${VERSION}, Zorro Version ${ZORROVERSION}"
 
 N="
@@ -118,8 +101,8 @@ echo 'FINISHED!'
 # clear | bash ./scripts/ci/build-delon.sh -n chart -debug
 if [[ ${DEBUG} == true ]]; then
   cd ../../
-  DEBUG_FROM=${PWD}/work/delon/dist/@delon/*
-  DEBUG_TO=${PWD}/work/ng-alain-themes/node_modules/@delon/
+  DEBUG_FROM=${PWD}/work/delon/dist/@delon/theme/*
+  DEBUG_TO=${PWD}/work/ng11-strict/node_modules/@delon/theme
   echo "DEBUG_FROM:${DEBUG_FROM}"
   echo "DEBUG_TO:${DEBUG_TO}"
   rm -rf ${DEBUG_TO}