+++ src/app/layout/layout.component.ts
@@ -1,8 +1,10 @@
+import { Location } from '@angular/common';
 import { Component, OnDestroy } from '@angular/core';
 import { NavigationEnd, NavigationError, RouteConfigLoadStart, Router } from '@angular/router';
+import { RTL, RTLService, SettingsService } from '@delon/theme';
 import { NzMessageService } from 'ng-zorro-antd/message';
 import { Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
+import { delay, filter, takeUntil } from 'rxjs/operators';
 
 @Component({
   selector: 'app-layout',
@@ -14,14 +16,21 @@ import { takeUntil } from 'rxjs/operators';
     <theme-btn></theme-btn>
   `,
   host: {
-    '[attr.id]': `'angular-content'`,
+    '[attr.id]': `'ng-content'`,
   },
 })
 export class LayoutComponent implements OnDestroy {
   private unsubscribe$ = new Subject<void>();
   isFetching = false;
 
-  constructor(router: Router, msg: NzMessageService) {
+  constructor(
+    private router: Router,
+    msg: NzMessageService,
+    private settingsSrv: SettingsService,
+    private location: Location,
+    rtl: RTLService,
+  ) {
+    rtl.change.subscribe(() => this.fixDirection());
     router.events.pipe(takeUntil(this.unsubscribe$)).subscribe(evt => {
       if (!this.isFetching && evt instanceof RouteConfigLoadStart) {
         this.isFetching = true;
@@ -36,6 +45,29 @@ export class LayoutComponent implements OnDestroy {
       }
       this.isFetching = false;
     });
+    router.events
+      .pipe(
+        takeUntil(this.unsubscribe$),
+        filter(ev => ev instanceof NavigationEnd),
+        delay(100),
+      )
+      .subscribe(() => {
+        this.fixDirection();
+      });
+  }
+
+  private fixDirection(): void {
+    // 修正rtl的query状态
+    const direction = this.settingsSrv.layout.direction;
+    const path = this.router.url.split(/[?#|?|#]/)[0];
+    const urlTree = this.router.parseUrl(this.router.url);
+    let fragment = urlTree.fragment;
+    if (fragment != null && fragment.length > 0) {
+      fragment = `#${fragment}`;
+    } else {
+      fragment = '';
+    }
+    this.location.replaceState(path, (direction === RTL ? `?direction=` + RTL : '') + fragment);
   }
 
   ngOnDestroy(): void {