+++ src/app/shared/shared-delon.module.ts
@@ -14,6 +14,7 @@ import { NoticeIconModule } from '@delon/abc/notice-icon';
 import { NumberToChineseModule } from '@delon/abc/number-to-chinese';
 import { OnboardingModule } from '@delon/abc/onboarding';
 import { PageHeaderModule } from '@delon/abc/page-header';
+import { PdfModule } from '@delon/abc/pdf';
 import { QRModule } from '@delon/abc/qr';
 import { QuickMenuModule } from '@delon/abc/quick-menu';
 import { ResultModule } from '@delon/abc/result';
@@ -39,6 +40,8 @@ import { G2TagCloudModule } from '@delon/chart/tag-cloud';
 import { G2TimelineModule } from '@delon/chart/timeline';
 import { TrendModule } from '@delon/chart/trend';
 import { G2WaterWaveModule } from '@delon/chart/water-wave';
+import { SettingDrawerModule } from '@delon/theme/setting-drawer';
+import { ThemeBtnModule } from '@delon/theme/theme-btn';
 
 export const SHARED_DELON_MODULES = [
   AvatarListModule,
@@ -83,4 +86,7 @@ export const SHARED_DELON_MODULES = [
   TrendModule,
   ReuseTabModule,
   MediaModule,
+  ThemeBtnModule,
+  SettingDrawerModule,
+  PdfModule,
 ];